requirements.txt
langchain-core
langchain
langchain-openai
python-dotenv
pydantic
langchain-community
gradio
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread